# docs/swagger.yaml
basePath: /v1/api
definitions:
...
  CONVERDA_internal_messaging_controller_dto.AddParticipantsRequest:
    properties:
      participants:
        items:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ParticipantDTO'
        minItems: 1
        type: array
    required:
    - participants
    type: object
  CONVERDA_internal_messaging_controller_dto.AgentDashboardResponse:
    properties:
      stats:
        $ref: '#/definitions/CONVERDA_internal_messaging_domain_model_entity.AgentStats'
    type: object
...
  CONVERDA_internal_messaging_controller_dto.AssignConversationRequest:
    properties:
      member_id:
        description: Optional, if empty assigns to caller
        type: string
    type: object
  CONVERDA_internal_messaging_controller_dto.BulkAssignRequest:
    properties:
      member_id:
        description: Optional
        type: string
      thread_ids:
        items:
          type: string
        type: array
    required:
    - thread_ids
    type: object
  CONVERDA_internal_messaging_controller_dto.CreateDirectChatRequest:
    properties:
      target:
        $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ParticipantDTO'
    required:
    - target
    type: object
  CONVERDA_internal_messaging_controller_dto.CreateGroupChatRequest:
    properties:
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ParticipantDTO'
        minItems: 1
        type: array
    required:
    - name
    - participants
    type: object
  CONVERDA_internal_messaging_controller_dto.InboundMessageRequest:
    properties:
      channel:
        description: Optional, default "api" or "web"
        type: string
      content:
        type: object
      environment_id:
        type: string
      subscriber_key:
        type: string
    required:
    - content
    - environment_id
    - subscriber_key
    type: object
  CONVERDA_internal_messaging_controller_dto.InternalNoteRequest:
    properties:
      content:
        type: object
    required:
    - content
    type: object
  CONVERDA_internal_messaging_controller_dto.MessageResponse:
    properties:
      content:
        type: object
      created_at:
        type: string
      id:
        type: string
      parent_id:
        type: string
      sender_type:
        type: string
      thread_id:
        type: string
    type: object
  CONVERDA_internal_messaging_controller_dto.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  CONVERDA_internal_messaging_controller_dto.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.Meta'
    type: object
  CONVERDA_internal_messaging_controller_dto.ParticipantDTO:
    properties:
      id:
        type: string
      type:
        enum:
        - user
        - subscriber
        type: string
    required:
    - id
    - type
    type: object
  CONVERDA_internal_messaging_controller_dto.ParticipantResponse:
    properties:
      display_name:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
    type: object
  CONVERDA_internal_messaging_controller_dto.PartnerDashboardResponse:
    properties:
      stats:
        allOf:
        - $ref: '#/definitions/CONVERDA_internal_messaging_domain_model_entity.TeamStats'
        description: Partner dashboard uses TeamStats for now
    type: object
  CONVERDA_internal_messaging_controller_dto.ReplyMessageRequest:
    properties:
      content:
        type: object
    required:
    - content
    type: object
  CONVERDA_internal_messaging_controller_dto.TeamDashboardResponse:
    properties:
      stats:
        $ref: '#/definitions/CONVERDA_internal_messaging_domain_model_entity.TeamStats'
    type: object
...
  CONVERDA_internal_messaging_controller_dto.ThreadResponse:
    properties:
      channel:
        type: string
      created_at:
        type: string
      environment_id:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      participants:
        items:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ParticipantResponse'
        type: array
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  CONVERDA_internal_messaging_controller_dto.UpdateGroupChatRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  CONVERDA_internal_messaging_domain_model_entity.AgentStats:
    properties:
      avg_response_time:
        type: number
      current_open_threads:
        type: integer
      member_id:
        type: string
      response_time_distribution:
        $ref: '#/definitions/CONVERDA_internal_messaging_domain_model_entity.ResponseTimeDistribution'
      total_assigned:
        type: integer
      total_resolved:
        type: integer
      workload:
        description: Same as CurrentOpenThreads for now
        type: integer
    type: object
  CONVERDA_internal_messaging_domain_model_entity.ResponseTimeDistribution:
    properties:
      from_5_to_15m:
        type: integer
      over_15m:
        type: integer
      under_5m:
        type: integer
    type: object
  CONVERDA_internal_messaging_domain_model_entity.TeamStats:
    properties:
      avg_response_time:
        type: number
      resolved_count:
        type: integer
      response_time_distribution:
        $ref: '#/definitions/CONVERDA_internal_messaging_domain_model_entity.ResponseTimeDistribution'
      sla_compliance_rate:
        type: number
      total_conversations:
        type: integer
    type: object
  CONVERDA_internal_notification_controller_dto.CreateGroupRequest:
    properties:
      description:
        type: string
      is_default:
        type: boolean
      key:
        type: string
      name:
        type: string
    required:
    - key
    - name
    type: object
...
  CONVERDA_internal_notification_controller_dto.GroupResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      key:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  CONVERDA_internal_notification_controller_dto.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  CONVERDA_internal_notification_controller_dto.UpdateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server for a Go Drunk Backend API, demonstrating DDD principles.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Drunk Backend API by DDD
  version: "1.0"
paths:
...
  /conversations:
    get:
      description: List conversations with filters
      parameters:
      - description: Status (unassigned, assigned, resolved)
        in: query
        name: status
        type: string
      - description: 'filter: ''me'' or ''all'''
        in: query
        name: assigned_to
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List Conversations
      tags:
      - Messaging
  /conversations/{id}:
    get:
      description: List messages in a conversation
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: string
      - description: Pagination Cursor (Base64)
        in: query
        name: cursor
        type: string
      - description: 'Direction: ''before'' (older) or ''after'' (newer)'
        in: query
        name: direction
        type: string
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Conversation Thread
      tags:
      - Messaging
  /conversations/{id}/assign:
    patch:
      consumes:
      - application/json
      description: Assign a conversation to an agent
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignee Data
        in: body
        name: request
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.AssignConversationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Assign Conversation
      tags:
      - Messaging
  /conversations/{id}/messages:
    post:
      consumes:
      - application/json
      description: Agent replies to a conversation pool
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: string
      - description: Message Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ReplyMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Reply to Conversation
      tags:
      - Messaging
  /conversations/{id}/notes:
    post:
      consumes:
      - application/json
      description: Add an internal note to a conversation
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      - description: Note Content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.InternalNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Add Internal Note
      tags:
      - Messaging
  /conversations/{id}/read:
    post:
      description: Mark all messages in thread as read
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Mark Thread Read
      tags:
      - Messaging
  /conversations/{id}/resolve:
    post:
      description: Mark conversation as resolved
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Resolve Conversation
      tags:
      - Messaging
  /conversations/{id}/unassign:
    post:
      description: Return conversation to the pool (unassigned)
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Unassign Conversation
      tags:
      - Messaging
  /conversations/assign:
    patch:
      consumes:
      - application/json
      description: Assign multiple conversations efficiently
      parameters:
      - description: Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.BulkAssignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Bulk Assign Conversations
      tags:
      - Messaging
  /conversations/direct:
    post:
      consumes:
      - application/json
      description: Create a 1-1 chat between 2 members or return existing one
      parameters:
      - description: Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.CreateDirectChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ThreadResponse'
      security:
      - BearerAuth: []
      summary: Create or Get Direct Chat
      tags:
      - Messaging
  /conversations/group:
    post:
      consumes:
      - application/json
      description: Create a global group chat
      parameters:
      - description: Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.CreateGroupChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.ThreadResponse'
      security:
      - BearerAuth: []
      summary: Create Group Chat
      tags:
      - Messaging
  /conversations/group/{id}:
    patch:
      consumes:
      - application/json
      description: Update group name
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.UpdateGroupChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update Group Chat
      tags:
      - Messaging
  /conversations/group/{id}/participants:
    post:
      consumes:
      - application/json
      description: Add members to group
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.AddParticipantsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Add Group Participants
      tags:
      - Messaging
  /conversations/group/{id}/participants/{memberID}:
    delete:
      description: Remove a member from group
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Remove Group Participant
      tags:
      - Messaging
  /conversations/inbound:
    post:
      consumes:
      - application/json
      description: Receive a message from external subscriber
      parameters:
      - description: Message Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.InboundMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.MessageResponse'
      summary: Receive Inbound Message
      tags:
      - Messaging
...
  /dashboards/agent/{member_id}:
    get:
      description: Get comprehensive agent performance metrics
      parameters:
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      - description: Environment ID
        in: query
        name: environment_id
        required: true
        type: string
      - description: From Time (RFC3339)
        in: query
        name: from
        type: string
      - description: To Time (RFC3339)
        in: query
        name: to
        type: string
      - description: SLA Threshold in seconds
        in: query
        name: sla_threshold
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.AgentDashboardResponse'
      security:
      - BearerAuth: []
      summary: Get Agent Dashboard
      tags:
      - Dashboards
  /dashboards/partner:
    get:
      description: Get comprehensive partner performance metrics
      parameters:
      - description: Environment ID
        in: query
        name: environment_id
        required: true
        type: string
      - description: From Time (RFC3339)
        in: query
        name: from
        type: string
      - description: To Time (RFC3339)
        in: query
        name: to
        type: string
      - description: SLA Threshold in seconds
        in: query
        name: sla_threshold
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.PartnerDashboardResponse'
      security:
      - BearerAuth: []
      summary: Get Partner Dashboard
      tags:
      - Dashboards
  /dashboards/team:
    get:
      description: Get comprehensive team performance metrics
      parameters:
      - description: Environment ID
        in: query
        name: environment_id
        required: true
        type: string
      - description: From Time (RFC3339)
        in: query
        name: from
        type: string
      - description: To Time (RFC3339)
        in: query
        name: to
        type: string
      - description: SLA Threshold in seconds
        in: query
        name: sla_threshold
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_messaging_controller_dto.TeamDashboardResponse'
      security:
      - BearerAuth: []
      summary: Get Team Dashboard
      tags:
      - Dashboards
...
  /notifications/groups:
    get:
      description: List notification groups
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List Notification Groups
      tags:
      - Notification
    post:
      consumes:
      - application/json
      description: Create a new notification group (category)
      parameters:
      - description: Group Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.GroupResponse'
      security:
      - BearerAuth: []
      summary: Create Notification Group
      tags:
      - Notification
  /notifications/groups/{id}:
    delete:
      description: Delete a notification group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Delete Notification Group
      tags:
      - Notification
    get:
      description: Get a notification group by ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.GroupResponse'
      security:
      - BearerAuth: []
      summary: Get Notification Group
      tags:
      - Notification
    patch:
      consumes:
      - application/json
      description: Update a notification group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CONVERDA_internal_notification_controller_dto.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update Notification Group
      tags:
      - Notification
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"